+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0 (2024-11-18)
+
+- The interface `IpAddrValue` now exposes an `octets()` method that returns a `Uint8Array`.
+- To prevent internal mutability of `IPv4Addr` and `IPv6Addr`, `IPv4Addr.octets` and `IPv6Addr.segments` are now readonly.
+
 ## 0.1.1 (2024-11-18)
 
 - The parser for IPv4 addresses and socket addresses no longer takes as many ASCII digits as possible.